deleted legacy files, reloacted files, working on cleanup
