+++ freecad/trails/project/commands/edit_alignment_cmd.py
@@ -31,28 +31,19 @@ import Draft
 import DraftTools
 from DraftTools import DraftTool, Creator
 from DraftTrackers import Tracker
+
 from ..support import utils, const
 from ..support.utils import Constants as C
 from ... import resources
 from ...alignment import horizontal_alignment as hz_align
-from ..tasks.alignment.draft_alignment_task import DraftAlignmentTask
-from . import wire_tracker
+#from ..tasks.alignment.draft_alignment_task import DraftAlignmentTask
+from ..trackers import wire_tracker
+from ..tasks.alignment import edit_alignment_task 
 
 class EditAlignmentCmd(DraftTool):
     """
     Initiates and manages drawing activities for alignment creation
     """
-
-    class STYLES(const.Const):
-        """
-        Internal constants used to define ViewObject styles
-        """
-
-        DISABLED =  [(0.4, 0.4, 0.4), 'Solid']
-        ENABLED =   [(0.8, 0.8, 0.8), 'Solid']
-        HIGHLIGHT = [(0.0, 1.0, 0.0), 'Solid']
-        PI =        [(0.0, 0.0, 1.0), 'Solid']
-        SELECTED =  [(1.0, 0.8, 0.0), 'Solid']
 
     def __init__(self):
         """
@@ -61,20 +52,9 @@ class EditAlignmentCmd(DraftTool):
 
         self.doc = None
 
-        self.view_objects = {
-            'selectables': [],
-            'line_colors': []
-        }
-
-        self.tmp_group = None
-        self.alignment = None
+        self.edit_alignment_task = None
         self.is_activated = False
-        self.event_cb = None
-        self.trackers = {}
-        self.active_tracker = None
-        self.selected_trackers = []
-        self.drag_tracker = None
-        self.radius_tracker = None
+        self.call = None
 
         self.button_states = {
             'BUTTON1': False,
@@ -129,52 +109,13 @@ class EditAlignmentCmd(DraftTool):
         self.doc = App.ActiveDocument
         self.view = Gui.ActiveDocument.ActiveView
 
-        #set up callback for SoEvents
-        self.event_cb = self.view.addEventCallback("SoEvent", self.action)
+        self.call = self.view.addEventCallback('SoEvent', self.action)
 
-        #disable selection entirely
-        self.view.getSceneGraph().getField("selectionRole").setValue(0)
-
-        #get all objects with LineColor and set them all to gray
-        self.view_objects['line_colors'] = \
-            [
-                (_v.ViewObject, _v.ViewObject.LineColor)
-                for _v in self.doc.findObjects()
-                if hasattr(_v, 'ViewObject')
-                if hasattr(_v.ViewObject, 'LineColor')
-            ]
-
-        for _v in self.view_objects['line_colors']:
-            self.set_style(_v[0], self.STYLES.DISABLED)
-
-        #create temporary group
-        self.tmp_group = self.doc.addObject('App::DocumentObjectGroup', 'Temp')
-
-        #create working, non-visual copy of horizontal alignment
-        data = Gui.Selection.getSelection()[0].Proxy.get_data_copy()
-        self.alignment = hz_align.create(data, utils.get_uuid(), True)
-
-        #deselect existing selections
-        Gui.Selection.clearSelection()
-
-        pi_points = self.alignment.get_pi_coords()
-
-        #create PI wire tracker geometry
-        self.wire_tracker = wire_tracker.create(
-            self.doc, 'Sugar_Grove_Rd_Horiz', pi_points
-        )
-
-        self.tmp_group.addObject(self.alignment.Object)
-
-        panel = DraftAlignmentTask(self.clean_up)
-
-        Gui.Control.showDialog(panel)
-        #panel.setup()
+        #create alignment editing task
+        self.edit_alignment_task = \
+            edit_alignment_task.create(self.doc, self.view)
 
         self.is_activated = True
-
-        self.doc.recompute()
-        DraftTools.redraw3DView()
 
     def action(self, arg):
         """
@@ -184,76 +125,17 @@ class EditAlignmentCmd(DraftTool):
         #trap the escape key to quit
         if arg['Type'] == 'SoKeyboardEvent':
             if arg['Key'] == 'ESCAPE':
+
                 self.finish()
                 return
 
         #trap mouse movement
         if arg['Type'] == 'SoLocation2Event':
-            
-            _p = self.view.getCursorPos()
-            _pt = self.view.getPoint(_p)
-            info = self.view.getObjectInfo(_p)
-
-            if info:
-                print(info)
-
-            return
-
-            if not self.button_states['BUTTON1']:
-                self.handle_mouseover_tracker(info)
-            else:
-                self.handle_drag_tracker(
-                    self.view.getPoint(_p)
-                )
+            pass
 
         #trap button clicks
         elif arg['Type'] == 'SoMouseButtonEvent':
-
-            self.get_button_states(arg)
-
-            _p = self.view.getCursorPos()
-            info = self.view.getObjectInfo(_p)
-
-            multi_select = arg['AltDown']
-
-            _key, _it = self.get_current_tracker(info)
-
-            #empty the arry and reset the trackers
-            if self.selected_trackers:
-
-                for _tracker in self.selected_trackers:
-                    _tracker.set_style(self.STYLES.ENABLED)
-
-                self.selected_trackers = []
-
-            #current defined?  add it to the selection.
-            if _it:
-
-                #Clear the button states so dragging won't work unles
-                for _v in self.button_states.values():
-                    _v = False
-
-                while True:
-
-                    _current = self.trackers[_key]
-                    _current.set_style(self.STYLES.SELECTED)
-
-                    self.active_tracker = _current
-                    self.selected_trackers.append(_current)
-
-                    if not multi_select:
-                        break
-
-                    try:
-                        _key = next(_it)
-
-                    except StopIteration:
-                        break
-
-                
-
-        self.doc.recompute()
-        DraftTools.redraw3DView()
+            pass
 
     def handle_drag_tracker(self, point):
         """
@@ -336,14 +218,6 @@ class EditAlignmentCmd(DraftTool):
 
         return None, None
 
-    def set_style(self, vobj, style):
-        """
-        Set the view object style based on the passed style tuple
-        """
-
-        vobj.LineColor = style[0]
-        vobj.DrawStyle = style[1]
-
     def clean_up(self):
         """
         Callback to finish the command
@@ -358,39 +232,39 @@ class EditAlignmentCmd(DraftTool):
         Finish drawing the alignment object
         """
 
+        if self.edit_alignment_task:
+            self.edit_alignment_task.finish()
+            self.edit_alignment_task = None
+
+        if self.call:
+            self.view.removeEventCallback('SoEvent', self.action)
+
+
+        #self.tmp_group.removeObjectsFromDocument()
+        #self.doc.removeObject(self.tmp_group.Name)
+
         #finalize tracking
         #if self.ui:
             #if hasattr(self, 'alignment_tracker'):
                 #self.alignment_tracker.finalize()
 
-        #delete the temproary geometry
-        self.tmp_group.removeObjectsFromDocument()
-        self.doc.removeObject(self.tmp_group.Name)
 
         #reset selctable object values
-        for _v in self.view_objects['selectables']:
-            _v.Selectable = True
+        #for _v in self.view_objects['selectables']:
+        #    _v.Selectable = True
 
         #reset line colors
-        for _v in self.view_objects['line_colors']:
-            _v[0].LineColor = _v[1]
+        #for _v in self.view_objects['line_colors']:
+        #    _v[0].LineColor = _v[1]
 
         #close dialog
-        if not Draft.getParam('UiMode', 1):
-            Gui.Control.closeDialog()
-
-        #remove the callback for action
-        if self.call:
-            self.view.removeEventCallback("SoEvent", self.call)
-
-        #shut down tracker
-        if self.wire_tracker:
-            self.wire_tracker.finalize()
+        #if not Draft.getParam('UiMode', 1):
+        #    Gui.Control.closeDialog()
 
         #re-enable selection
-        self.view.getSceneGraph().getField("selectionRole").setValue(1)
+        #self.view.getSceneGraph().getField("selectionRole").setValue(1)
 
-        Creator.finish(self)
+       # Creator.finish(self)
 
         self.is_activated = False
